@malcolmbaig 2021-01-20
Some comments left inline, but overall some really nice improvements here 👍🏽